✨ refactor(repository): update save method to accept Partial<Task> and add NotFoundException handling in delete method

+++ src/task/repository/impl/task.repository.ts
@@ -1,7 +1,7 @@
 import { InjectModel } from '@nestjs/mongoose';
 import { ITaskRepository } from '../task-repository.interface';
 import { Task } from 'src/task/schemas/task.schema';
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { Model } from 'mongoose';
 
 @Injectable()
@@ -10,7 +10,7 @@ export class TaskRepository implements ITaskRepository {
     @InjectModel(Task.name) private readonly taskModel: Model<Task>,
   ) {}
 
-  async save(task: Task): Promise<Task> {
+  async save(task: Partial<Task>): Promise<Task> {
     const newTask = new this.taskModel(task);
     return await newTask.save();
   }
@@ -28,6 +28,10 @@ export class TaskRepository implements ITaskRepository {
   }
 
   async delete(id: string): Promise<boolean> {
+    const task = await this.taskModel.findById(id).exec();
+    if (!task) {
+      throw new NotFoundException('Tarea no encontrada');
+    }
     const result = await this.taskModel.deleteOne({ _id: id }).exec();
     return result.deletedCount === 1;
   }
